requirements.txt
arrow==1.2.3
cachetools==5.2.0
certifi==2024.7.4
charset-normalizer==2.0.12
DateTime==4.7
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.6.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
icalendar==5.0.1
icalevents==0.1.25
idna==3.7
oauthlib==3.2.2
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.5
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.19
zope.interface==5.5.0
